app/requirements.txt
faster_whisper==1.0.3
Flask==3.0.3
Flask_Cors==4.0.1
python-dotenv==1.0.1
Requests==2.32.3
colorlog==6.8.2
gunicorn==22.0.0